tests: Move channel creation tests to test_channel_creation.
